handle dead citation links gracefully with an error message
When we render section previews, simple-2 generates dead links for
internal citations because we do not have an index file available
that maps section citations to their filename. It puts a call to
a Javascript function instead.

This commit implements that javascript function. Otherwise it would
log an error to the console about the function not being defined.